refactor: code splitting on `GroupInfo`
