+++ app/(authorized)/chats/group_info/GroupInfo.tsx
@@ -2,30 +2,14 @@
 
 import React, { useEffect, useState } from "react";
 import { Swiper, SwiperSlide } from "swiper/react";
-import ImagePreviewModal from "../[group]/ImagePreviewModal";
 import {
     AccordionDetails,
     Box,
     Button,
     ButtonGroup,
-    Dialog,
-    DialogActions,
-    DialogContent,
-    DialogContentText,
-    DialogTitle,
-    FilledInput,
-    FormControl,
-    IconButton,
-    ImageList,
-    ImageListItem,
-    InputAdornment,
-    InputLabel,
-    OutlinedInput,
-    Tooltip,
     Typography,
 } from "@mui/material";
 import HoverWrapper from "../../HoverWrapper";
-import { LuCopy } from "react-icons/lu";
 import SharedLink from "./SharedLink";
 import { Swiper as SwiperType } from "swiper/types";
 import { Accordion, AccordionSummary } from "./GroupInfoHelpers";
@@ -33,21 +17,19 @@ import { MemberTile } from "./MemberTile";
 import EditMemberNicknameTile from "./EditMemberNicknameTile";
 import ChangeGroupEmoji from "./ChangeGroupEmoji";
 import { useToastAlert } from "@/contexts/ToastAlertContext";
-import {
-    changeGroupInformation,
-    sendInfoMessageToGroup,
-} from "@/lib/helpers/firebase-helpers";
-import { auth, firestoreDb } from "@/firebase";
+import { auth } from "@/firebase";
 import { useRouter } from "next/navigation";
-import { DATABASE_PATH, ROLE, SITEMAP } from "@/lib/variables";
+import { ROLE } from "@/lib/variables";
 import AddMembersAlertDialog from "./AddMembersAlertDialog";
 import { IGroupDetails, MessageType } from "@/app";
 import { URL_REGEX } from "@/lib/utils";
-import { LoadingButton } from "@mui/lab";
 import VoiceMessageBox from "../[group]/VoiceMessageBox";
 import EditAndDisplayAvatar from "./EditAndDisplayAvatar";
 import { useUsers } from "@/contexts/UsersInGroupContext";
-import { arrayRemove, doc, updateDoc } from "firebase/firestore";
+import GroupLink from "./GroupLink";
+import GroupAdminAccordion from "./GroupAdminAccordion";
+import ImageSwiperSlide from "./ImageSwiperSlide";
+import LeaveGroup from "./LeaveGroup";
 
 export default function GroupInfo({
     messages = [],
@@ -58,12 +40,10 @@ export default function GroupInfo({
 }) {
     const [activeTab, setActiveTab] = useState<0 | 1 | 2 | number>(0);
     const [swiper, setSwiper] = useState<SwiperType>();
-    const [showImageModal, setShowImageModal] = useState("");
     const [openLeaveDialog, setOpenLeaveDialog] = useState(false);
     const [addUserDialog, setAddUserDialog] = useState(false);
     const { allUsers } = useUsers();
     const [loading, setLoading] = useState(false);
-    const [groupname, setGroupname] = useState(group?.name || "");
     const [myRole, setMyRole] = useState<number | undefined>();
     const handleTabChange = (newValue: number) => {
         setActiveTab(newValue);
@@ -86,14 +66,6 @@ export default function GroupInfo({
         );
     }, [allUsers]);
 
-    let imageLink = messages
-        .map((msg) => msg.imageLink)
-        .flat(Infinity) as string[];
-
-    const handleClose = () => {
-        setShowImageModal("");
-    };
-
     return (
         <>
             <Box className="w-full overflow-y-auto h-full">
@@ -105,53 +77,22 @@ export default function GroupInfo({
                 >
                     {group?.name || "Group name"}
                 </Typography>
-                <HoverWrapper className="mx-4 w-[calc(100%-2.1rem)]">
-                    <FormControl variant="filled" fullWidth>
-                        <InputLabel
-                            htmlFor="filled-adornment-password"
-                            className="font-bold"
-                        >
-                            Group Link
-                        </InputLabel>
-                        <FilledInput
-                            value={
-                                !!location
-                                    ? location.origin + group?.inviteLink
-                                    : group?.inviteLink
-                            }
-                            disabled
-                            fullWidth
-                            endAdornment={
-                                <InputAdornment position="end">
-                                    <Tooltip title="Copy">
-                                        <IconButton
-                                            aria-label="toggle password visibility"
-                                            onClick={() => {
-                                                navigator.clipboard.writeText(
-                                                    !!location
-                                                        ? location.origin +
-                                                              (group?.inviteLink ||
-                                                                  "")
-                                                        : group?.inviteLink ||
-                                                              ""
-                                                );
-                                            }}
-                                            edge="end"
-                                        >
-                                            <LuCopy />
-                                        </IconButton>
-                                    </Tooltip>
-                                </InputAdornment>
-                            }
-                        />
-                    </FormControl>
-                </HoverWrapper>
+                <GroupLink inviteLink={group?.inviteLink || ""} />
                 <Accordion className="mt-4">
                     <AccordionSummary>Change Group Emoji</AccordionSummary>
                     <AccordionDetails>
                         <ChangeGroupEmoji />
                     </AccordionDetails>
                 </Accordion>
+                {/* ADMIN PANEL */}
+                {(group?.groupMembersBasicDetails.find(
+                    (e) => e.id === auth.currentUser?.uid
+                )?.role ?? ROLE.member) < ROLE.member && (
+                    <Accordion elevation={3}>
+                        <AccordionSummary>✨ Manage Group ✨</AccordionSummary>
+                        <GroupAdminAccordion loading={loading} />
+                    </Accordion>
+                )}
                 <Accordion>
                     <AccordionSummary>See Members</AccordionSummary>
                     <AccordionDetails>
@@ -202,79 +143,10 @@ export default function GroupInfo({
                         ))}
                     </AccordionDetails>
                 </Accordion>
-
-                {/* ADMIN PANEL */}
-                {(group?.groupMembersBasicDetails.find(
-                    (e) => e.id === auth.currentUser?.uid
-                )?.role ?? ROLE.member) < ROLE.member && (
-                    <Accordion>
-                        <AccordionSummary>
-                            Manage Group (Only admins or higher)
-                        </AccordionSummary>
-                        <AccordionDetails>
-                            <FormControl
-                                variant="outlined"
-                                fullWidth
-                                sx={{
-                                    gridArea: "edit",
-                                }}
-                            >
-                                <OutlinedInput
-                                    fullWidth
-                                    defaultValue={group?.name}
-                                    // value={groupname}
-                                    onChange={(e) => {
-                                        setGroupname(e.target.value);
-                                    }}
-                                    endAdornment={
-                                        <InputAdornment position="end">
-                                            <LoadingButton
-                                                size="small"
-                                                variant="contained"
-                                                loading={loading}
-                                                sx={{
-                                                    fontSize: ".70rem",
-                                                }}
-                                                onClick={async () => {
-                                                    if (!group)
-                                                        return setMessage(
-                                                            "Warning: Group is not resolved"
-                                                        );
-                                                    if (groupname.length < 3) {
-                                                        return setMessage(
-                                                            "Error: cancelling operation. Name must be at least 3 characters long"
-                                                        );
-                                                    }
-                                                    if (groupname.length > 20) {
-                                                        return setMessage(
-                                                            "Error: cancelling operation. Name must be less than 20 characters long"
-                                                        );
-                                                    }
-                                                    if (
-                                                        groupname === group.name
-                                                    ) {
-                                                        return setMessage(
-                                                            "Error: group name didn't change"
-                                                        );
-                                                    }
-
-                                                    changeGroupInformation(
-                                                        group.id,
-                                                        {
-                                                            name: groupname,
-                                                        }
-                                                    );
-                                                }}
-                                            >
-                                                <span>set</span>
-                                            </LoadingButton>
-                                        </InputAdornment>
-                                    }
-                                />
-                            </FormControl>
-                        </AccordionDetails>
-                    </Accordion>
-                )}
+                <Accordion>
+                    <AccordionSummary>Pinned Message</AccordionSummary>
+                    <AccordionDetails></AccordionDetails>
+                </Accordion>
 
                 <Accordion className="mb-4">
                     <AccordionSummary>Media and files</AccordionSummary>
@@ -325,41 +197,7 @@ export default function GroupInfo({
                             }}
                         >
                             <SwiperSlide>
-                                <HoverWrapper className="w-[100%-1rem]">
-                                    <ImageList
-                                        variant="masonry"
-                                        cols={3}
-                                        gap={8}
-                                        className="my-0 rounded-[inherit] w-[100%-1rem]"
-                                    >
-                                        {imageLink.map((src, i) => {
-                                            return (
-                                                <ImageListItem
-                                                    key={i}
-                                                    component="a"
-                                                    href={`#${i + 1}`}
-                                                    onClick={() => {
-                                                        setShowImageModal(
-                                                            i.toString()
-                                                        );
-                                                    }}
-                                                >
-                                                    <img
-                                                        src={`${src}?w=248&fit=crop&auto=format`}
-                                                        srcSet={`${src}?w=248&fit=crop&auto=format&dpr=2 2x`}
-                                                        alt={src}
-                                                        loading="lazy"
-                                                    />
-                                                </ImageListItem>
-                                            );
-                                        })}
-                                        <ImagePreviewModal
-                                            images={imageLink}
-                                            handleClose={handleClose}
-                                            showImageModal={showImageModal}
-                                        />
-                                    </ImageList>
-                                </HoverWrapper>
+                                <ImageSwiperSlide />
                             </SwiperSlide>
                             <SwiperSlide>
                                 <Box>
@@ -407,91 +245,7 @@ export default function GroupInfo({
                         </Swiper>
                     </AccordionDetails>
                 </Accordion>
-                <Button
-                    color="error"
-                    variant="contained"
-                    className="block max-w-[600px] mx-auto my-4"
-                    sx={{ mx: "auto", display: `block` }}
-                    fullWidth
-                    onClick={() => {
-                        setOpenLeaveDialog(true);
-                    }}
-                >
-                    Leave Group
-                </Button>
-                <Dialog
-                    open={openLeaveDialog}
-                    onClose={handleCloseLeaveDialog}
-                    aria-labelledby="alert-dialog-title"
-                    aria-describedby="alert-dialog-description"
-                >
-                    <DialogTitle>Leave Group?</DialogTitle>
-                    <DialogContent>
-                        <DialogContentText>
-                            Are you sure to leave{" "}
-                            <Typography color="primary.main" component={"span"}>
-                                {group?.name || "<group-name>"}
-                            </Typography>{" "}
-                            group?
-                        </DialogContentText>
-                    </DialogContent>
-                    <DialogActions>
-                        <LoadingButton
-                            loading={loading}
-                            onClick={async () => {
-                                setLoading(true);
-                                try {
-                                    if (!group) return;
-                                    if (!auth.currentUser) return;
-
-                                    const myInfoOnGroup =
-                                        group.groupMembersBasicDetails.find(
-                                            (e) =>
-                                                e.id === auth.currentUser?.uid
-                                        );
-
-                                    if (!myInfoOnGroup) {
-                                        return setMessage(
-                                            "Error: something went wrong. please reLogin"
-                                        );
-                                    }
-                                    await sendInfoMessageToGroup(
-                                        `${myInfoOnGroup.nickname} [${myInfoOnGroup.id}] has left the group`,
-                                        group.id,
-                                        myInfoOnGroup
-                                    );
-
-                                    await updateDoc(
-                                        doc(
-                                            firestoreDb,
-                                            DATABASE_PATH.groupDetails,
-                                            group.id
-                                        ),
-                                        {
-                                            groupMembers: arrayRemove(
-                                                auth.currentUser?.uid
-                                            ),
-                                        }
-                                    );
-                                    setMessage("You left from group.");
-                                    push(SITEMAP.chats);
-                                } catch (e) {
-                                    setMessage("Error: check console");
-                                    console.log(e);
-                                }
-                                setLoading(false);
-                                handleCloseLeaveDialog();
-                            }}
-                            variant="contained"
-                            color="error"
-                        >
-                            <span>Yes</span>
-                        </LoadingButton>
-                        <Button onClick={handleCloseLeaveDialog} autoFocus>
-                            No
-                        </Button>
-                    </DialogActions>
-                </Dialog>
+                <LeaveGroup />
             </Box>
         </>
     );
